pyproject.toml
[project]
name = "clipboard-qr"
version = "0.1.0"
description = "Recognize QrCode from your clipboard."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow>=11.3.0",
    "pyperclip>=1.11.0",
    "pywin32>=311 ; sys_platform == 'win32'",
    "pyzbar>=0.1.9",
    "qrcode>=8.2",
]

[project.scripts]
qr = "main:main"
